requirements.txt
click>=8.0.3,<8.1
pydantic>=1.8.2,<1.9
PyYAML>=6.0,<6.1
typer>=0.4.0,<0.5
typing-extensions>=4.0.1,<4.1
rich>=12.0.0,<12.1
GitPython>=3.1.27,<3.2
